risk-agents-app/backend/pyproject.toml
# Risk Agents Backend - Python Project Configuration
# This file defines the project and its dependencies (managed by UV)

[project]
name = "risk-agents-backend"
version = "0.1.0"
description = "Risk Agents AI-powered project management backend"
readme = "README.md"
requires-python = ">=3.11"

# Core dependencies
dependencies = [
    "fastapi>=0.104.0",           # Modern web framework for building APIs
    "uvicorn[standard]>=0.24.0",  # ASGI server to run FastAPI
    "pydantic>=2.0.0",            # Data validation using Python type hints
    "python-jose[cryptography]>=3.3.0",  # JWT token handling
    "python-multipart>=0.0.6",    # Form data parsing
    "pyyaml>=6.0",                # YAML parsing for skills
    "anthropic>=0.34.0",          # Anthropic SDK
    "passlib[bcrypt]>=1.7.4",     # Password hashing with bcrypt (Module 3.2)
    "argon2-cffi>=21.3.0",        # Argon2 password hashing backend
    "python-dotenv>=1.0.0",       # Environment variable loading
]

[project.optional-dependencies]
# Development dependencies (testing, linting, etc.)
dev = [
    "pytest>=7.4.0",              # Testing framework
    "pytest-asyncio>=0.21.0",     # Async testing support
    "black>=23.0.0",              # Code formatter
    "ruff>=0.1.0",                # Fast Python linter
    "httpx>=0.25.0",              # HTTP client for testing
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "agent"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
